+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Add optional .onion block explorer links for BTC/TBTC in advanced settings
+- Android: show onion/clearnet chooser when opening block explorer links with .onion URLs enabled
 - Bitcoin: show account details for the persisted receive address type by default
 - Add external block explorer links to used addresses
 - Settings search